% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JM_imp_adaptive.R
\name{as_inits}
\alias{as_inits}
\title{Return the current state of a 'JointAI' model as list to be used as
initial values}
\usage{
as_inits(object, what = c("RinvD", "invD", "tau", "b"))
}
\arguments{
\item{object}{an object of class 'JointAI'}

\item{what}{vector of node types}
}
\description{
Return the current state of a 'JointAI' model as list to be used as
initial values
}
