% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_missingness.R
\name{create_missingness}
\alias{create_missingness}
\title{Create missingness in a \code{data.frame}}
\usage{
create_missingness(
  data,
  mis_scen,
  varlvls = NULL,
  groups = NULL,
  idvars = NULL,
  mis_seed = NULL
)
}
\arguments{
\item{data}{a \code{data.frame}}

\item{mis_scen}{a missingness scenario}

\item{varlvls}{named vector giving the level of each variable}

\item{groups}{list with grouping information}

\item{idvars}{optional vector with names of ID variables that should be used
to identify levels in a multi-level setting}

\item{mis_seed}{optional seed value}
}
\description{
Create missingness in a \code{data.frame}
}
