% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_missingness.R
\name{create_missings}
\alias{create_missings}
\title{#' Create missingness in a \code{data.frame}
#' @inheritParams delete_MCAR
#' @param mis_scen a missingness scenario
create_missingness <- function(data, mis_scen, varlvls = NULL,
groups = NULL, idvars = NULL, mis_seed = NULL) {
if (is.null(mis_scen)) {
return(data)
} else {
args <- c(list(data = data),
mis_scen$fun_args,
idvars = idvars,
if (is.null(mis_scen$fun_args$varlvls)) {
list(varlvls = varlvls)
},
if (is.null(mis_scen$fun_args$groups)) {
list(groups = groups)
},
if (is.null(mis_scen$fun_args$mis_seed)) {
c(mis_seed = mis_seed)
})
do.call(mis_scen$fun, args)
}
}}
\usage{
create_missings(
  data,
  idvars,
  prop_mis,
  prop_cc,
  patterns = NULL,
  weights = NULL,
  type = "RIGHT",
  groups = NULL,
  varlvls = NULL
)
}
\arguments{
\item{data}{a complete \code{data.frame}}

\item{prop_mis}{a named vector indicating the proportion of missing values
per variable (only variables with proportions > 0 have to be
included)}

\item{prop_cc}{the proportion of complete cases}

\item{patterns}{a matrix specifying the missing data patterns to be used; a
\code{1} indicates a variable is completely observed and \code{0} that
it is incomplete; if unspecified, this is created
automatically.}

\item{weights}{a matrix of the same dimension as \code{patterns} that indicates
which variables are predictive of missingness in this pattern;
this argument allows to define MCAR, MAR or MNAR missingness.
If not specified, \code{weights} is set equal to \code{patterns}, i.e.,
MAR with all observed variables being predictors.}

\item{type}{A string or vector of strings containing the type of missingness for each
pattern. Either \code{"LEFT"}, \code{"MID"}, \code{"TAIL"} or '\code{"RIGHT"}.
If a single missingness type is given, all patterns will be created with the same
type. If the missingness types should differ between patterns, a vector of missingness
types should be given. Default is RIGHT for all patterns and is the result of
\code{\link[mice]{ampute.default.type}}.}
}
\description{
Create missing values in a \code{data.frame}
}
