% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_JointAI.R
\name{create_simulation_scenario}
\alias{create_simulation_scenario}
\title{Create a simulation scenario from a JointAI object}
\usage{
create_simulation_scenario(object, file)
}
\arguments{
\item{object}{an object of class "JointAI"}

\item{file}{path and file name where the extracted information should be
saved to.}
}
\description{
Extracts parameters needed to simulate covariates (\code{extract_covar_pars()})
and the parameters needed to simulate the corresponding outcome(s)
(\code{extract_outcome_pars()}).
}
