% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_MAR.R
\name{delete_MAR}
\alias{delete_MAR}
\title{Create MAR missingness}
\usage{
delete_MAR(
  data,
  missing,
  idvars = NULL,
  varlvls = NULL,
  groups = NULL,
  seed = NULL
)
}
\arguments{
\item{data}{a \code{data.frame}}

\item{missing}{a named vector giving the names of covariates and proportion
of values to be deleted}

\item{idvars}{optional vector with names of ID variables that should be used
to identify levels in a multi-level setting}

\item{varlvls}{named vector giving the level of each variable}

\item{groups}{list with grouping information}
}
\description{
Create MAR missingness
}
