% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_MCAR.R
\name{delete_MCAR}
\alias{delete_MCAR}
\title{Create MCAR missingness}
\usage{
delete_MCAR(
  data,
  missing,
  idvars = NULL,
  varlvls = NULL,
  groups = NULL,
  mis_seed = NULL
)
}
\arguments{
\item{data}{a \code{data.frame}}

\item{missing}{a named vector giving the names of covariates and proportion
of values to be deleted}

\item{idvars}{optional vector with names of ID variables that should be used
to identify levels in a multi-level setting}

\item{varlvls}{named vector giving the level of each variable}

\item{groups}{list with grouping information}

\item{mis_seed}{optional seed value}
}
\description{
Create MCAR missingness
}
