% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_JointAI.R
\name{extract_basehaz}
\alias{extract_basehaz}
\title{Extract the posterior mean of the regression coefficients for a spline
specification of the baseline hazard}
\usage{
extract_basehaz(object)
}
\arguments{
\item{object}{object of class 'JointAI'}
}
\description{
Extract the posterior mean of the regression coefficients for a spline
specification of the baseline hazard
}
