% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_JointAI.R
\name{extract_basehaz_knots}
\alias{extract_basehaz_knots}
\title{Extract the knots used for the spline specification of the baseline hazard
from a model of class 'JointAI'}
\usage{
extract_basehaz_knots(object)
}
\arguments{
\item{object}{object of class 'JointAI'}
}
\description{
Extract the knots used for the spline specification of the baseline hazard
from a model of class 'JointAI'
}
