% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_JointAI.R
\name{extract_mean_cens}
\alias{extract_mean_cens}
\title{Extract the mean censoring time form an object of class 'JointAI'}
\usage{
extract_mean_cens(object)
}
\arguments{
\item{object}{object of class 'JointAI'}
}
\description{
Extract the mean censoring time form an object of class 'JointAI'
}
