% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_JointAI.R
\name{extract_outcome_pars}
\alias{extract_outcome_pars}
\title{Extract parameters necessary to simulate an outomce from an object of class
'JointAI'}
\usage{
extract_outcome_pars(object)
}
\arguments{
\item{object}{object of class 'JointAI'}
}
\description{
This function obtains the \code{response_type} (from the \code{analysis_type}),
\code{formula}, \code{reg_coefs} (\code{extract_reg_coefs()}), \code{type}s (\code{extract_types()}),
baseline hazard coefficients and knots (\code{extract_basehaz()} and
\code{extract_basehaz_knots()}), mean censoring time (\code{extract_mean_cens()}),
upper limit of where to search for the event time (\code{.up}) and the step
size with which to increase that upper limit (\code{up_step}).
}
