% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_JointAI.R
\name{extract_rd_vcov}
\alias{extract_rd_vcov}
\title{Extract the posterior mean of the random effects variance-covariance matrices
from an object of class 'JointAI'}
\usage{
extract_rd_vcov(object)
}
\arguments{
\item{object}{object of class 'JointAI'}
}
\description{
Extract the posterior mean of the random effects variance-covariance matrices
from an object of class 'JointAI'
}
