% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_JointAI.R
\name{extract_reg_coefs}
\alias{extract_reg_coefs}
\title{Extract the posterior mean of the regression coefficients from an object of
class 'JointAI'}
\usage{
extract_reg_coefs(object)
}
\arguments{
\item{object}{object of class 'JointAI'}
}
\description{
Extract the posterior mean of the regression coefficients from an object of
class 'JointAI'
}
