% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_JointAI.R
\name{extract_resid_sd}
\alias{extract_resid_sd}
\title{Extract the posterior mean of the residual standard deviations from an object
of class 'JointAI'}
\usage{
extract_resid_sd(object)
}
\arguments{
\item{object}{object of class 'JointAI'}
}
\description{
Extract the posterior mean of the residual standard deviations from an object
of class 'JointAI'
}
