% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_JointAI.R
\name{extract_types}
\alias{extract_types}
\title{Extract the model types of longitudinal variables from an object of class
'JointAI'}
\usage{
extract_types(object)
}
\arguments{
\item{object}{object of class 'JointAI'}
}
\description{
Extract the model types of longitudinal variables from an object of class
'JointAI'
}
