% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_models}
\alias{fit_models}
\title{Fit one more multiple models and obtain the model result}
\usage{
fit_models(models, formula, data, seed = NULL, scen = NULL)
}
\arguments{
\item{models}{a list of model information}

\item{formula}{the model formula}

\item{data}{a \code{data.frame}}

\item{seed}{an optional seed value}

\item{scen}{optional; name of scenario}
}
\description{
Fit one more multiple models and obtain the model result
}
