% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_info.R
\name{get_compl_data_info}
\alias{get_compl_data_info}
\title{Summarize information of the complete simulated data}
\usage{
get_compl_data_info(data, data_seed, idvars, data_lvls, ...)
}
\arguments{
\item{data}{a \code{data.frame}}

\item{data_seed}{the seed value used to generate the data}

\item{idvars}{optional vector with names of ID variables that should be used
to identify levels in a multi-level setting}

\item{data_lvls}{named vector giving the level of each variable}

\item{...}{optional additional arguments}
}
\description{
Summarize information of the complete simulated data
}
