% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_info.R
\name{get_data_info}
\alias{get_data_info}
\title{Summarize information of the simulated data}
\usage{
get_data_info(data, seed, ...)
}
\arguments{
\item{data}{a \code{data.frame}}

\item{seed}{the seed value}

\item{...}{optional additional arguments}
}
\description{
Summarize information of the simulated data
}
