% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JM_imp_adaptive.R
\name{get_inits_JM}
\alias{get_inits_JM}
\title{Generate initial values for a JointAI joint model}
\usage{
get_inits_JM(formula, data, inits_iter, n_chains, seed = NULL)
}
\arguments{
\item{formula}{the model formula}

\item{data}{the dataset}

\item{inits_iter}{number of iterations}

\item{n_chains}{number of chains (has to be the same as in the joint model)}

\item{seed}{seed value}
}
\description{
Generate initial values for a JointAI joint model
}
