% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_info.R
\name{get_miss_data_info}
\alias{get_miss_data_info}
\title{Summarize information of the incomplete simulated data}
\usage{
get_miss_data_info(data, seed, idvars, data_lvls, scen = NULL, ...)
}
\arguments{
\item{data}{a \code{data.frame}}

\item{seed}{the seed value}

\item{idvars}{optional vector with names of ID variables that should be used
to identify levels in a multi-level setting}

\item{data_lvls}{named vector giving the level of each variable}

\item{scen}{optional, name of the missingness scenario}

\item{...}{optional additional arguments; not used}
}
\description{
Summarize information of the incomplete simulated data
}
