% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_functions.R
\name{get_res_df}
\alias{get_res_df}
\title{Convert the simulation results into a data.frame
The resulting \code{data.frame} will contain the simulation results from all
models, the true regression coefficients (\code{regcoef}) and simulation
evaluation measures (bias, relative bias, coverage of the CI, CI width and
MSE) per simulated dataset.}
\usage{
get_res_df(object)
}
\arguments{
\item{object}{object of class simulation_result}
}
\description{
Convert the simulation results into a data.frame
The resulting \code{data.frame} will contain the simulation results from all
models, the true regression coefficients (\code{regcoef}) and simulation
evaluation measures (bias, relative bias, coverage of the CI, CI width and
MSE) per simulated dataset.
}
