% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_result_.R
\name{get_result_default}
\alias{get_result_default}
\title{Extract a summary of the model results from a standard model
This function uses \code{coef()} and \code{confint()}.}
\usage{
get_result_default(fitted_model, type = NA, seed = NA, ...)
}
\arguments{
\item{fitted_model}{a fitted model object}

\item{type}{optional character string specifying the type (to identify
what type of model the results belong to later in the visualization
and description of results)}

\item{seed}{optional (but) suggested seed value. Will be used when called
from within \code{run_models()}.}

\item{...}{optional additional arguments for compatibility with other
\verb{get_result_<...>} functions}
}
\description{
Extract a summary of the model results from a standard model
This function uses \code{coef()} and \code{confint()}.
}
