% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_result_JointAI}
\alias{get_result_JointAI}
\title{Extract a summary of the model results from a JointAI model
This function uses \code{coef()} and \code{confint()}.}
\usage{
get_result_JointAI(fitted_model, seed = NA, outcome = 1L, ...)
}
\arguments{
\item{fitted_model}{a fitted model object}

\item{seed}{optional (but) suggested seed value. Will be used when called
from within \code{run_models()}.}

\item{outcome}{integer identifying for which outcome the results should be
extracted}

\item{...}{optional additional arguments for compatibility with other
\verb{get_result_<...>} functions}
}
\description{
Extract a summary of the model results from a JointAI model
This function uses \code{coef()} and \code{confint()}.
}
