% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_result_.R
\name{get_result_lme4}
\alias{get_result_lme4}
\title{Extract a summary of the model results from a lme4 model
This function uses \code{fixef()} and \code{confint()}.}
\usage{
get_result_lme4(
  fitted_model,
  type = NA,
  seed = NA,
  scen = NA,
  ci_method = "Wald",
  ...
)
}
\arguments{
\item{fitted_model}{a fitted model object}

\item{type}{optional character string specifying the type (to identify
what type of model the results belong to later in the visualization
and description of results)}

\item{seed}{optional (but) suggested seed value. Will be used when called
from within \code{run_models()}.}

\item{scen}{optional name of scenario}

\item{ci_method}{method passed to \code{confint()}}

\item{...}{optional additional arguments for compatibility with other
\verb{get_result_<...>} functions}
}
\description{
Extract a summary of the model results from a lme4 model
This function uses \code{fixef()} and \code{confint()}.
}
