% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_simulation.R
\name{make_mv_vcov}
\alias{make_mv_vcov}
\title{Combine multiple variance-covariance matrices to one multivariate matrix}
\usage{
make_mv_vcov(..., structure = "indep")
}
\arguments{
\item{...}{separate variance-covariance matrices}

\item{structure}{character string indicating the correlation structure
between the matrices (at the moment only "indep")}
}
\description{
Combine multiple variance-covariance matrices to one multivariate matrix
}
\examples{
make_mv_vcov(
make_vcov(rgamma(2, 3, 2), sim_corr_unif(2)),
make_vcov(rgamma(4, 3, 2), sim_corr_unif(4))
)

}
