% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_simulation.R
\name{make_vcov}
\alias{make_vcov}
\title{Create a variance-covariance matrix from a vector of variances and a
vector of correlations}
\usage{
make_vcov(variances, correlations)
}
\arguments{
\item{variances}{numeric vector}

\item{correlations}{numeric vector}
}
\description{
Create a variance-covariance matrix from a vector of variances and a
vector of correlations
}
