% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_check.R
\name{run_jointai_flex}
\alias{run_jointai_flex}
\title{Run a JointAI model with flexible number of iterations}
\usage{
run_jointai_flex(fun, formula, data, model_args, gr_check_args, ...)
}
\arguments{
\item{fun}{JointAI model function}

\item{model_args}{a list of arguments passed to \code{fun}}

\item{gr_check_args}{a list of arguments to be passed to the gr check}
}
\description{
Run a JointAI model with flexible number of iterations
}
