% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{run_sim}
\alias{run_sim}
\title{Run simulation}
\usage{
run_sim(
  sim_pars,
  covar_def,
  outcome_pars,
  models,
  mis_scenarios = NULL,
  sim_name = NULL,
  scenarios = NULL,
  batch_size = 10,
  batch_numbers = NULL,
  path = NULL,
  packages = NULL,
  skip_fit = FALSE
)
}
\arguments{
\item{sim_pars}{list of simulation parameters}

\item{covar_def}{expression defining covariates}

\item{outcome_pars}{list of parameters to create covariates and outcome}

\item{models}{list of model specifications}

\item{mis_scenarios}{list of missingness scenarios}

\item{sim_name}{optional character string used in naming the output folder}

\item{batch_size}{optional; number of simulations per batch (will be saved
in separate file)}

\item{batch_numbers}{optional index numbers of the batches of simulations to be run
(in order to only run a subset)}

\item{path}{where to save the results (file name will be generated
automatically)}

\item{packages}{optional character vector of packages passed to \code{.packages}
in \code{foreach::foreach}}

\item{skip_fit}{logical; should the step in which the models are fitted be
skipped (to only get the simulated data)?}
}
\description{
Run simulation
}
