% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R
\name{set_scenarios}
\alias{set_scenarios}
\title{Create a data.frame of the scenarios to run.
A scenario consists of the missing data scenario and model to be run on this
data.}
\usage{
set_scenarios(miss_scenarios, models)
}
\arguments{
\item{miss_scenario}{a named list of missingness scenarios}

\item{model}{a named list of model specifications}
}
\description{
Create a data.frame of the scenarios to run.
A scenario consists of the missing data scenario and model to be run on this
data.
}
