% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions_simulation.R
\name{sim_corr_unif}
\alias{sim_corr_unif}
\title{Simulate uniformly distributed correlation elements for a variance-covariance matrix}
\usage{
sim_corr_unif(dim, min = -0.5, max = 0.5)
}
\arguments{
\item{dim}{the dimension of the matrix}

\item{min}{minimum used in \code{runif()}}

\item{max}{maximum used in \code{runif()}}
}
\description{
Simulate uniformly distributed correlation elements for a variance-covariance matrix
}
