% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_covars_mvn.R
\name{sim_covars_mvn}
\alias{sim_covars_mvn}
\title{Simulate data from a multivariate normal distribution}
\usage{
sim_covars_mvn(N, covar_pars, seed = NULL, ...)
}
\arguments{
\item{N}{the sample size}

\item{covar_pars}{a list with objects \code{means}, \code{vcov}, \code{probs} and \code{ordered}}

\item{seed}{seed value}

\item{...}{additional parameters (not used)}
}
\description{
Data are simulated from a multivariate normal distribution. To create
factors, the columns for which \code{probs} and \code{ordered}is specified are
converted into factors by cutting the simulated continuous variable at the
quantiles given by \code{probs}.
}
