% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{Simulate data}
\usage{
sim_data(covar_def, data_params, seed = NULL)
}
\arguments{
\item{covar_def}{covariate definition}

\item{data_params}{a list of parameters}

\item{seed}{a seed value}
}
\description{
Simulate data
}
