% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_outcome_glmm.R
\name{sim_outcome_glmm}
\alias{sim_outcome_glmm}
\title{Simulate a (multivariate) longitudinal outcome}
\usage{
sim_outcome_glmm(
  data,
  formula,
  reg_coefs,
  resid_sd = NULL,
  ranef_vcov,
  type = "gaussian",
  return_ranefs = FALSE,
  ranefs = NULL,
  ...
)
}
\arguments{
\item{data}{a \code{data.frame} containing covariate data}

\item{formula}{a list of model formulas for the longitudinal outcomes}

\item{reg_coefs}{a named list of regression coefficients for
each model, with names equal to the corresponding response
variable. Each of the list elements is a (named) vector.
If named, the names will be matched with the names of the
design matrix that is created from \code{formula} and \code{data}.}

\item{resid_sd}{named vector of residual standard deviations, with names
equal to the corresponding response variable (for types
"gaussian" and "Gamma")}

\item{ranef_vcov}{a random effects variance-covariance matrix or a named
list of such matrices when there are more than two levels
(and names being equal to the grouping variables)}

\item{type}{named vector of model types. Available model types are
"gaussian", "binomial", "Gamma", and "poisson".}

\item{...}{arguments passed to other functions}

\item{seed}{the seed value}
}
\description{
Simulate a (multivariate) longitudinal outcome
}
\examples{
# Bivariate outcome in a multi-level setting


covar_data <- function(N, J) {
  merge(
    # subject level information
    data.frame(id = 1:N,
               center = rep(1:J, each = N/2),
               age = rnorm(N, 60, 10),
               sex = factor(rbinom(N, size = 1, prob = 0.5))),
    # repeated measurements
    data.frame(id = rep(1:N, each = J),
               time = runif(J * N, 0, 5)
    )
  )
}

ranef_vcov <- list(id = Matrix::bdiag(
  matrix(nrow = 2, ncol = 2, data = c(2, 0.5, 0.5, 1),
         dimnames = list(c('int', 'time'), c('int', 'time'))),
  matrix(nrow = 3, ncol = 3, data = c(  2,  0.5,  0.1,
                                        0.5, 1.5, -0.2,
                                        0.1, -0.2,  1),
         dimnames = list(c('int', 'time1', "time2"), c('int', 'time1', 'time2')))
),
center = matrix(nrow = 2, ncol = 2, data = 0.2,
                dimnames = list(c('int', 'int'), c('int', 'int')))
)


formula <- list(x1 ~ age + sex + time + (time | id) + (1 | center),
                x2 ~ age + x1 + sex + ns(time, df = 2) + (ns(time, df = 2) | id) +
                  (1 | center)
)

reg_coefs = list(x1 = c(1:4),
                 x2 = c(-100,  2, -0.01, -3, -2, -1))
sim_outcome_glmm(data = covar_data(N = 50, J = 2),
                 formula = formula,
                 reg_coefs = reg_coefs,
                 resid_sd = c(x1 = 0.5),
                 ranef_vcov = ranef_vcov,
                 type = c(x1 = "gaussian", x2 = "binomial"))
}
