% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_outcome_lm.R
\name{sim_outcome_lm}
\alias{sim_outcome_lm}
\title{Simulate a conditionally normally distributed outcome}
\usage{
sim_outcome_lm(data, formula, reg_coefs, resid_sd, seed = NULL, ...)
}
\arguments{
\item{data}{a \code{data.frame} containing covariate data}

\item{formula}{the model formula}

\item{reg_coefs}{a vector of regression coefficients. If named, the names
will be matched with the names of the design matrix that
is created from \code{formula} and \code{data}.}

\item{resid_sd}{residual standard deviation}

\item{seed}{optional seed value}

\item{...}{other arguments passed to other functions}
}
\description{
Simulate a conditionally normally distributed outcome
}
