% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_outcome_jm.R
\name{sim_outcome_surv_jm}
\alias{sim_outcome_surv_jm}
\title{Simulate a survival and longitudinal outcome for a joint model}
\usage{
sim_outcome_surv_jm(
  data,
  formula,
  reg_coefs,
  resid_sd,
  beta_Bh0 = NULL,
  ranef_vcov,
  type,
  timevar = "time",
  basehaz_type = "splines",
  phi = 1.2,
  mean_cens = 30,
  .tries = 5,
  .up = 500L,
  up_step = 500L,
  seed = NULL,
  no_obs_after_event = TRUE,
  ...
)
}
\arguments{
\item{data}{a \code{data.frame} containing covariate data}

\item{formula}{a list of model formulas for the survival and longitudinal
outcomes}

\item{reg_coefs}{a list of regression coefficients with elements "surv" and
"long". Each of the list elements is a (named) vector.
If named, the names will be matched with the names of the
design matrix that is created from \code{formula} and \code{data}.}

\item{ranef_vcov}{a random effects variance-covariance matrix or a named
list of such matrices when there are more than two levels
(and names being equal to the grouping variables)}

\item{phi}{parameter of the Wishart distribution}

\item{mean_cens}{mean censoring time}

\item{.tries}{integer; how often is the upper limit increased when looking
for the root}

\item{seed}{the seed value}

\item{...}{arguments passed to other functions}
}
\description{
Simulate a survival and longitudinal outcome for a joint model
}
