% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simvalidator.R
\docType{package}
\name{simvalidator}
\alias{simvalidator}
\title{Simvalidator: validation of model performance via simulation}
\description{
Functions to run simulations and summarize and visualize the results
}
