% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_JointAI.R
\name{extract_covar_pars}
\alias{extract_covar_pars}
\title{Extract parameters necessary to simulate covariate data from an object of
class 'JointAI'}
\usage{
extract_covar_pars(object)
}
\arguments{
\item{object}{object of class 'JointAI'}
}
\description{
Extract parameters necessary to simulate covariate data from an object of
class 'JointAI'
}
