% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_result_.R
\name{get_result_JointAI}
\alias{get_result_JointAI}
\title{Extract a summary of the model results from a JointAI model
This function uses \code{coef()} and \code{confint()}.}
\usage{
get_result_JointAI(
  fitted_model,
  data_seed = NA,
  outcome = 1L,
  subset = NULL,
  miss_scenario = NA,
  model = NULL,
  ...
)
}
\arguments{
\item{fitted_model}{a fitted model object}

\item{data_seed}{optional (but) suggested seed value. Will be used when called
from within \code{run_models()}.}

\item{outcome}{integer identifying for which outcome the results should be
extracted}

\item{subset}{subset specification of \code{JointAI::coef()},
\code{JointAI::confint()}, \code{JointAI::GR_crit} and
\code{JointAI::MC_error()}}

\item{miss_scenario}{optional name of missingness scenario}

\item{model}{optional name for model to be included in summary}

\item{...}{optional additional arguments for compatibility with other
\verb{get_result_<...>} functions}
}
\description{
Extract a summary of the model results from a JointAI model
This function uses \code{coef()} and \code{confint()}.
}
