% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{run_sim}
\alias{run_sim}
\title{Run simulation}
\usage{
run_sim(sim_params, covar_def, data_params, models, path = NULL)
}
\arguments{
\item{sim_params}{list of simulation parameters}

\item{covar_def}{expression defining covariates}

\item{data_params}{list of parameters to create covariates and outcome}

\item{models}{list of model specifications}

\item{path}{where to save the results (file name will be generated
automatically)}
}
\description{
Run simulation
}
