% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{run_sim}
\alias{run_sim}
\title{Run simulation}
\usage{
run_sim(
  sim_pars,
  covar_def,
  outcome_pars,
  models,
  path = NULL,
  packages = NULL
)
}
\arguments{
\item{sim_pars}{list of simulation parameters}

\item{covar_def}{expression defining covariates}

\item{outcome_pars}{list of parameters to create covariates and outcome}

\item{models}{list of model specifications}

\item{path}{where to save the results (file name will be generated
automatically)}

\item{packages}{optional character vector of packages passed to \code{.packages}
in \code{foreach::foreach}}
}
\description{
Run simulation
}
