% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_outcome_glmm.R
\name{sim_outcome_glmm}
\alias{sim_outcome_glmm}
\title{Simulate a (multivariate) longitudinal outcome}
\usage{
sim_outcome_glmm(
  data,
  formula,
  reg_coefs,
  resid_sd = NULL,
  ranef_vcov,
  type = "gaussian",
  return_ranefs = FALSE,
  seed = NULL,
  ...
)
}
\arguments{
\item{data}{a \code{data.frame} containing covariate data}

\item{formula}{a list of model formulas for the longitudinal outcomes}

\item{reg_coefs}{a named list of regression coefficients for
each model, with names equal to the corresponding response
variable. Each of the list elements is a (named) vector.
If named, the names will be matched with the names of the
design matrix that is created from \code{formula} and \code{data}.}

\item{resid_sd}{named vector of residual standard deviations, with names
equal to the corresponding response variable (for types
"gaussian" and "Gamma")}

\item{ranef_vcov}{a random effects variance-covariance matrix or a named
list of such matrices when there are more than two levels
(and names being equal to the grouping variables)}

\item{type}{named vector of model types. Available model types are
"gaussian", "binomial", "Gamma", and "poisson".}

\item{return_ranefs}{logical; should the random effects be returned?
If \code{TRUE}, a list with \code{data} and \code{ranefs} is returned.}

\item{seed}{the seed value}

\item{...}{arguments passed to other functions}
}
\description{
Simulate a (multivariate) longitudinal outcome
}
\examples{
# Bivariate outcome in a multi-level setting






}
