% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_outcome_survival.R
\name{sim_outcome_survival}
\alias{sim_outcome_survival}
\title{Simulate a survival outcome}
\usage{
sim_outcome_survival(
  data,
  formula,
  reg_coefs,
  phi = 1.2,
  mean_cens = 30,
  .tries = 5L,
  seed = NULL,
  ...
)
}
\arguments{
\item{data}{a \code{data.frame} containing covariate data}

\item{formula}{the model formula}

\item{reg_coefs}{a vector of regression coefficients. If named, the names
will be matched with the names of the design matrix that
is created from \code{formula} and \code{data}.}

\item{phi}{parameter of the Wishart distribution}

\item{mean_cens}{mean censoring time}

\item{.tries}{integer; how often is the upper limit increased when looking
for the root}

\item{seed}{the seed value}

\item{...}{arguments passed to other functions}
}
\description{
Simulate a survival outcome
}
